Revert "Add language routing to internal navigation in SPA (#3503)"
This reverts commit ffd2a79fd1116d9d22b7a1a78d39069ed248295d.
