+++ locale-handler.js
@@ -63,7 +63,7 @@ module.exports = (req, res, next) => {
 
   if (lang === locale.defaultLocale) {
     // If this is alread default language, continue.
-    return isLangInPath ? res.redirect(path.join('/', filePath)) : next();
+    return next();
   }
 
   const localizedFilePath = path.join(
@@ -76,6 +76,6 @@ module.exports = (req, res, next) => {
   if (fs.existsSync(localizedFilePath)) {
     return isLangInPath ? next() : res.redirect(path.join('/', lang, filePath));
   } else {
-    return isLangInPath ? res.redirect(path.join('/', filePath)) : next();
+    return res.redirect(path.join('/', locale.defaultLocale, filePath));
   }
 };
