+++ src/lib/app.js
@@ -12,7 +12,6 @@ import '@webcomponents/webcomponentsjs/webcomponents-loader.js';
 import './analytics'; // side effects
 import {swapContent, getPartial} from './loader';
 import * as router from './utils/router';
-import {checkUserPreferredLanguage} from './actions';
 import {store} from './store';
 import {localStorage} from './utils/storage';
 import removeServiceWorkers from './utils/sw-remove';
@@ -37,9 +36,6 @@ WebComponents.waitFor(async () => {
   });
 });
 
-// Read preferred language from the url, a cookie or browser settings.
-checkUserPreferredLanguage();
-
 // Configures global page state (loading, signed in).
 function onGlobalStateChanged({isSignedIn, isPageLoading}) {
   document.body.classList.toggle('lh-signedin', isSignedIn);
@@ -53,6 +49,7 @@ function onGlobalStateChanged({isSignedIn, isPageLoading}) {
   } else {
     main.removeAttribute('aria-busy');
   }
+
   // Cache whether the user was signed in, to help prevent FOUC in future and
   // for Analytics, as this can be read synchronosly and Firebase's auth takes
   // ~ms to arrive.
