+++ src/lib/store.js
@@ -36,7 +36,6 @@ const initialState = {
   lighthouseError: null,
 
   currentUrl: window.location.pathname,
-  currentLanguage: document.documentElement.getAttribute('lang'),
   isOffline: Boolean(getMeta('offline')),
   isSideNavExpanded: false,
   isModalOpen: false,
@@ -55,7 +54,7 @@ const initialState = {
   showingSnackbar: false,
   snackbarType: null,
 
-  userPreferredLanguage: '',
+  userPrefferedLanguage: '',
 
   // Used to override the current time for web.dev/LIVE testing.
   timeOffset,
