+++ src/lib/utils/language.js
@@ -14,12 +14,6 @@ const languageNames = {
 };
 
 /**
- * A default language for the site.
- * @const
- */
-const defaultLanguage = 'en';
-
-/**
  * A list of supported languages.
  * @const
  */
@@ -34,15 +28,8 @@ function isValidLanguage(lang) {
   return supportedLanguages.indexOf(lang) > -1;
 }
 
-function getLanguageFromPath(path) {
-  const parts = path.split('/');
-  return isValidLanguage(parts[1]) && parts[1];
-}
-
 export default {
   languageNames,
-  defaultLanguage,
   isValidLanguage,
   supportedLanguages,
-  getLanguageFromPath,
 };
