+++ src/site/_includes/partials/post.njk
@@ -32,6 +32,8 @@
     {% include 'partials/subscribe-action.njk' %}
   </div>
 
+  {% include 'partials/translation-disclaimer.njk' %}
+
   <div class="w-layout-container--narrow w-post-content">
     <header class="w-article-header">
       {% if 'blog' in tags %}
