+++ src/site/content/en/en.11tydata.js
@@ -20,11 +20,9 @@ module.exports = function () {
   ].filter(livePaths);
 
   const lang = 'en';
-  const locale = 'en_US';
 
   return {
     lang,
-    locale,
     home: {
       paths,
     },
